Decouple ANT Runtime API
