+++ lite/test/build-ant-option2.sh
@@ -4,6 +4,7 @@ MODULE_PATH=../../lite/api-modules/
 cd ../../dep/iotjs/
 ./tools/build.py --target-board=rpi3 \
     --cmake-param=-DENABLE_MODULE_ANT=ON \
+    --cmake-param=-DENABLE_MODUKE_ANTRUNTIME=ON \
     --cmake-param=-DENABLE_MODULE_ANTCOMPANION=ON \
     --cmake-param=-DENABLE_MODULE_ANTRESOURCE=ON \
     --cmake-param=-DENABLE_MODULE_ANTREMOTEUI=ON \
