+++ lite/test/build-ant-tx2.sh
@@ -4,11 +4,12 @@ MODULE_PATH=../../lite/api-modules/
 cd ../../dep/iotjs/
 ./tools/build.py --target-board=tx2 \
     --cmake-param=-DENABLE_MODULE_ANT=ON \
+    --cmake-param=-DENABLE_MODUKE_ANTRUNTIME=ON \
     --cmake-param=-DENABLE_MODULE_ANTCOMPANION=ON \
-    --cmake-param=-DENABLE_MODULE_ANTML=ON \
-    --cmake-param=-DENABLE_MODULE_ANTSTREAM=ON \
-    --cmake-param=-DENABLE_MODULE_ANTREMOTEUI=ON \
     --cmake-param=-DENABLE_MODULE_ANTRESOURCE=ON \
+    --cmake-param=-DENABLE_MODULE_ANTREMOTEUI=ON \
+    --cmake-param=-DENABLE_MODULE_ANTSTREAM=ON \
+    --cmake-param=-DENABLE_MODULE_ANTML=ON \
     --external-modules=${MODULE_PATH}
 cd -
